commented out prints

+++ src/main.cpp
@@ -138,7 +138,7 @@ void drive_forward_inches(long inches){
 void publishEncCounts(){
     int ecounts = EC.get_encoder_counts();
     char result[20];
-    enc_val.data = (float)ecounts;
+    enc_val.data = ecounts;
     dtostrf(ecounts, 20, 5, result);
     pub_enc.publish(enc_val);
     nh.logwarn(result);
